Schematics: add pressure sensors
The design allows 2 pressure sensors: 'sensor 1' and 'sensor 2'.

Only sensor 1 is required. Sensor 2 can be left out, but may be placed in case of
changing requirements in the future (e.g. using 2 sensors to measure
flow).

Each of the sensors may be implemented by choosing one of 3 different footprints.
On the PCB layout, these 3 footprints will be overlapped to create a
'universal' pressure sensor footprint.